- add the animation of throwing a character, and the impact animation when the throwed character touches the ground, with the impact sound. - correct the size of the firefall and of their collision box.

+++ raw-asset/items/generatePNGs.py
@@ -4,7 +4,7 @@ from PIL import Image
 
 listfile = glob.glob("./*.png")
 for filename in listfile:
-    if "crown" not in filename:
+    if "crown" not in filename and "ground_shock" not in filename:
         os.remove(filename)
 
 output_path = ".\\"
